requirements.txt
tensorflow
numpy
Pillow
streamlit
matplotlib